+++ mikeplus/__init__.py
@@ -1,19 +1,53 @@
 """MIKE+Py package."""
 
-__version__ = "2025.1.0"
+__version__ = "2025.6.0"
 
-from .conflicts import check_conflicts
 
-check_conflicts()
+from pathlib import Path
 
+from .conflicts import check_conflicts as _check_conflicts
+
+_check_conflicts()
+
+from pythonnet import load  # noqa: E402
+
+load(
+    "coreclr",
+    runtime_config=(
+        (Path(__file__).parent / "bin" / "runtimeconfig.json").absolute().as_posix()
+    ),
+)
 import clr  # noqa: E402
 
-clr.AddReference(
-    "DHI.Mike.Install, Version=1.0.0.0, Culture=neutral, PublicKeyToken=c513450b5d0bf0bf"
-)
-from DHI.Mike.Install import MikeImport, MikeProducts  # noqa: E402
+from .utils import setup_bin_path as _setup_bin_path  # noqa: E402
 
-MikeImport.Setup(23, MikeProducts.MikePlus)
+_install_root, _dll_dir_handle = _setup_bin_path(
+    major_assembly_version=23,
+    fallback_mikeplus_install_root=Path("C:/Program Files (x86)/DHI/MIKE+/2025"),
+    env_var_name_install_root="MIKEPLUSPY_INSTALL_ROOT",  # set this environment variable to use custom install path
+    bin_path=Path("bin/x64"),
+)
+
+#  keep here for backward compatibility (mikeio1d uses) ... remove in 2026.0.0
+try:
+    from DHI.Mike.Install import (
+        MikeImport,  # noqa: E402, F401
+        MikeProducts,  # noqa: E402, F401
+    )
+except ImportError:
+    # mock this case:
+    # mikeplus.MikeImport.ActiveProduct().InstallRoot
+    # used by mikeio1d
+    class _MockMikeImport:  # noqa: E402, F401
+        @staticmethod
+        def ActiveProduct():
+            return _MockMikeProduct()
+
+    class _MockMikeProduct:  # noqa: E402, F401
+        InstallRoot = _install_root
+
+    MikeImport = _MockMikeImport
+    MikeProduct = _MockMikeProduct
 
 clr.AddReference("System")
 clr.AddReference("System.Runtime")
@@ -25,11 +59,15 @@ clr.AddReference("DHI.Amelia.DataModule.Interface")
 clr.AddReference("DHI.Amelia.Infrastructure.Interface")
 clr.AddReference("DHI.Amelia.GlobalUtility")
 clr.AddReference("DHI.Amelia.Tools.EngineTool")
+clr.AddReference("DHI.Amelia.EPANETBridge")
+clr.AddReference("DHI.Amelia.SWMMBridge")
 
-from .datatableaccess import DataTableAccess  # noqa: E402
-from .datatableaccess import DataTableDemoAccess  # noqa: E402
 from .database import Database  # noqa: E402
-from .shortcuts import open, create  # noqa: E402
+from .datatableaccess import (  # noqa: E402
+    DataTableAccess,  # noqa: E402
+    DataTableDemoAccess,  # noqa: E402
+)
+from .shortcuts import create, open  # noqa: E402
 from .utils import to_sql  # noqa: E402
 
 __all__ = ["Database", "open", "create", "to_sql"]